pack up client as a module
